feat(client): added css to header
