+++ src/App.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { BrowserRouter, Route } from 'react-router-dom';
 import LandingPage from './views/landing/LandingPage';
 import SignUp from './views/signup/SignUp';
-import Header from './views/header_footer/Header';
-import Footer from './views/header_footer/Footer';
+import Header from './~reusables/layout/Header';
+import Footer from './~reusables/layout/Footer';
 
 function App() {
 	return (
