+++ src/~reusables/variables/FontSizes.js
@@ -1,0 +1,6 @@
+//====== FONT SIZES ======//
+export const heading_1 = '48px';
+export const heading_2 = '32px';
+export const heading_3 = '24px';
+export const heading_4 = '16px';
+export const heading_5 = '12px';
